Added tests of CNTK on scrambled images
Testing for circulant structure of NTK under arbitrary permutations.
We know that the CNTK produces circulant stuff if given shifted images.
The reason is that those shifts "preserve" the information.
Here we try to see what happens if we play around with different, general
permutation matrices, with and without GlobalAvgPool. Having it or not results
in different reduction strategies for the computation of the actual kernel.
In general, we find that the circulant nature is not preserved for scrambled
images under CNN, while it is under MLP (makes sense, as MLPs are permutation
invariant, and arbitrary scrambles in general break local structure).